requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
requests==2.32.3
numpy==1.24.3
pandas==2.0.3
polygon-api-client==1.14.0
websocket-client==1.6.1
python-dotenv==1.0.1
gunicorn==21.2.0
